Move popular_activity_json out of view and into script.
